[char_property] Create module dir and add submodules
Put the contract traits and macro all next to each other under one
module dir.

Also update some documentations.
