+++ unic/utils/src/lib.rs
@@ -29,7 +29,6 @@ pub const PKG_DESCRIPTION: &'static str = env!("CARGO_PKG_DESCRIPTION");
 
 pub mod char_property;
 pub mod codepoints;
-mod macros;
 pub mod tables;
